QObjects API

Index

SymBasis.DoFObjects.DoFObject

SymBasis.DoFObjects.bint

SymBasis.DoFObjects.dof_object

SymBasis.DoFObjects.dof_object

SymBasis.DoFObjects.DoFObject — Type

DoFObject{B,T_ldof,T<:Integer,Ti<:Integer} <: SymBasis.DoFObjects.AbstractDoFObject
DoFObject(
    type::Symbol,
    ldof::NTuple{B,T_ldof};
    T::DataType=UInt, Ti::DataType=Int
) where {B,T_ldof}

An object having degrees of freedom. Each object is characterized by its type and a tuple of local degrees of freedom (ldof) with length B. The type of the local degrees of freedom is specified by the type parameter T_ldof, while the integer types used for indexing and calculations are specified by the type parameters T and Ti.

Fields

type::Symbol: The type of the object (e.g., :Spin, :Fermion) as a symbol.

ldof::NTuple{B,T_ldof}: A tuple representing the local degrees of freedom of the object.

Constructor Arguments

type::Symbol: The type of the object (e.g., :Spin, :Fermion) as a symbol.

ldof::NTuple{B,T_ldof}: A tuple representing the local degrees of freedom of the object.

Constructor Keyword Arguments

T::DataType=UInt: The integer type used for indexing and calculations. Default is UInt.

Ti::DataType=Int: The integer type used for indexing and calculations. Default is Int.

Returns

DoFObject{B,T_ldof,T,Ti}: A new DoFObject instance.

SymBasis.DoFObjects.bint — Method

bint(
    dofo::SymBasis.DoFObjects.DoFObject{B,T_ldof,T,Ti}
) where {B,T_ldof,T<:Integer,Ti<:Integer}

Returns SymBasis.DigitBase.BaseInt corresponding to the object.

Arguments

dofo::SymBasis.DoFObjects.DoFObject{B,T_ldof,T,Ti}: The object for which to

create the base integer.

Returns

SymBasis.DigitBase.BaseInt{T,Ti,B}: The base integer type corresponding to the object.

SymBasis.DoFObjects.dof_object — Method

dof_object(sym::Symbol, args...; kwargs...)

Create an object with degrees of freedom based on the specified symbol sym and additional arguments.

Arguments

sym::Symbol: The symbol representing the type of object to create.

args...: Additional positional arguments required for the object constructor.

Keyword Arguments

kwargs...: Additional keyword arguments required for the object constructor.

Returns

SymBasis.DoFObjects.DoFObject: The created degree of freedom object.

SymBasis.DoFObjects.dof_object — Method

dof_object(type::Val{:Spin}, s::Rational; T::Type=UInt, Ti::Type=Int)

Create a quantum mechanical spin object with spin s.

Arguments

type::Val{:Spin}: The type of the object, fixed to :Spin.

s::Rational: The spin value.

T::Type=UInt: The underlying integer type for storage (default is UInt).

Ti::Type=Int: The integer type used for indexing (default is Int).

Returns

SymBasis.DoFObjects.DoFObject{B,T_ldof,T,Ti}: The object representing the spin.