SymGroups API

Index

SymBasis.SymGroups.CombSymGroup

SymBasis.SymGroups.SymGroup

Base.:∘

Base.:∘

Base.:∘

Base.:∘

SymBasis.SymGroups._check_Nₛ

SymBasis.SymGroups._check_Nₛ

SymBasis.SymGroups._make_hashset

SymBasis.SymGroups._make_hashset

SymBasis.SymGroups.apply_Nₛ

SymBasis.SymGroups.apply_perm

SymBasis.SymGroups.apply_perm

SymBasis.SymGroups.check_Nₛ

SymBasis.SymGroups.check_perm

SymBasis.SymGroups.sym

SymBasis.SymGroups.sym

SymBasis.SymGroups.sym

SymBasis.SymGroups.sym

SymBasis.SymGroups.CombSymGroup — Type

CombSymGroup{B,T_s,T<:Integer,Ti<:Integer,T_f<:Union{Number,Complex{<:Number}}}
CombSymGroup(
    dofo::DoFObject{B,T_s,T,Ti},
    cycles::AbstractArray{<:AbstractVector{<:NamedTuple}},
    check::AbstractVector{<:Function},
    apply::AbstractVector{<:Function},
    factors::AbstractArray{T_f},
    N::Integer
) where {B,T_s,T<:Integer,Ti<:Integer,T_f<:Union{Number,Complex{<:Number}}}

A combined symmetry group formed by the composition of multiple symmetry groups acting on the same DoF-object. This structure allows for the representation of more complex symmetry operations by combining simpler ones.

Fields

dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti}: The DoF-object on which the combined symmetry group acts.

cycles::AbstractArray{<:AbstractVector{<:NamedTuple}}: An array of vectors of named tuples representing the combined symmetry cycles.

check::AbstractVector{<:Function}: A vector of functions to check the validity of each set of symmetry operations.

apply::AbstractVector{<:Function}: A vector of functions to apply each set of symmetry operations.

factors::AbstractArray{T_f}: An array of factors associated with each combined symmetry cycle.

N::Integer: The total number of the DoF-objects in the system. This is used to check the validity of the symmetry operations.

Constructor Arguments

dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti}: The DoF-object on which the combined symmetry group acts.

cycles::AbstractArray{<:AbstractVector{<:NamedTuple}}: An array of vectors of named tuples representing the combined symmetry cycles.

check::AbstractVector{<:Function}: A vector of functions to check the validity of each set of symmetry operations.

apply::AbstractVector{<:Function}: A vector of functions to apply each set of symmetry operations.

factors::AbstractArray{T_f}: An array of factors associated with each combined symmetry cycle.

N::Integer: The total number of the DoF-objects in the system. This is used to check the validity of the symmetry operations.

Returns

CombSymGroup{B,T_s,T,Ti,T_f}: A new CombSymGroup instance initialized with the provided parameters.

The constructor checks that the size of cycles matches the size of factors and that the number of dimensions matches the number of check and apply functions to ensure consistency.

SymBasis.SymGroups.SymGroup — Type

SymGroup{B,T_s,T<:Integer,Ti<:Integer,T_f<:Union{Number,Complex{<:Number}}}
SymGroup(
    dofo::DoFObject{B,T_s,T,Ti},
    cycles::AbstractVector{<:NamedTuple},
    check::Function,
    apply::Function,
    factors::AbstractVector{T_f},
    N::Integer
) where {B,T_s,T<:Integer,Ti<:Integer,T_f<:Union{Number,Complex{<:Number}}}

A symmetry group acting on a DoF-object. The symmetry group is defined by its associated DoF-object, a set of cycles representing the symmetry operations, functions to check and apply these operations, and factors associated with each symmetry cycle.

Fields

dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti}: The DoF-object on which the symmetry group acts.

cycles::AbstractVector{<:NamedTuple}: A vector of named tuples representing the symmetry cycles.

check::Function: A function to check the validity of symmetry operations.

apply::Function: A function to apply the symmetry operations.

factors::AbstractVector{T_f}: A vector of factors associated with each symmetry cycle.

N::Integer: The total number of the DoF-objects in the system. This is used to check the validity of the symmetry operations.

Constructor Arguments

dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti}: The DoF-object on which the symmetry group acts.

cycles::AbstractVector{<:NamedTuple}: A vector of named tuples representing the symmetry cycles.

check::Function: A function to check the validity of symmetry operations.

apply::Function: A function to apply the symmetry operations.

factors::AbstractVector{T_f}: A vector of factors associated with each symmetry cycle.

N::Integer: The total number of the DoF-objects in the system. This is used to check the validity of the symmetry operations.

Returns

SymGroup{B,T_s,T,Ti,T_f}: A new SymGroup instance initialized with the provided parameters.

The constructor checks that the number of cycles matches the number of factors to ensure consistency.

Base.:∘ — Method

∘(
    csg1::CombSymGroup{B,T_s,T,Ti,T_f1},
    csg2::CombSymGroup{B,T_s,T,Ti,T_f2}
) where {
    B,
    T_s,
    T<:Integer,
    Ti<:Integer,
    T_f1<:Union{Number,Complex{<:Number}},
    T_f2<:Union{Number,Complex{<:Number}}
}

Composition of two combined symmetry groups acting on the same DoF-object. The resulting symmetry group combines the cycles, check functions, apply functions, and factors of the input symmetry groups.

Arguments

csg1::SymBasis.SymGroups.CombSymGroup{B,T_s,T,Ti,T_f1}: The first combined symmetry group.

csg2::SymBasis.SymGroups.CombSymGroup{B,T_s,T,Ti,T_f2}: The second combined symmetry group.

Returns

SymBasis.SymGroups.CombSymGroup{B,T_s,T,Ti,typeof(*(one(T_f1), one(T_f2)))}: The combined symmetry group.

Base.:∘ — Method

∘(
    csg::CombSymGroup{B,T_s,T,Ti,T_f1},
    sg::SymGroup{B,T_s,T,Ti,T_f2}
) where {
    B,
    T_s,
    T<:Integer,
    Ti<:Integer,
    T_f1<:Union{Number,Complex{<:Number}},
    T_f2<:Union{Number,Complex{<:Number}}
}

Composition of a combined symmetry group with a symmetry group acting on the same DoF-object. The resulting symmetry group combines the cycles, check functions, apply functions, and factors of the input symmetry groups.

Arguments

csg::SymBasis.SymGroups.CombSymGroup{B,T_s,T,Ti,T_f1}: The combined symmetry group.

sg::SymBasis.SymGroups.SymGroup{B,T_s,T,Ti,T_f2}: The symmetry group.

Returns

SymBasis.SymGroups.CombSymGroup{B,T_s,T,Ti,typeof(*(one(T_f1), one(T_f2)))}: The combined symmetry group.

Base.:∘ — Method

∘(
    csg::SymGroup{B,T_s,T,Ti,T_f1},
    sg::CombSymGroup{B,T_s,T,Ti,T_f2}
) where {
    B,
    T_s,
    T<:Integer,
    Ti<:Integer,
    T_f1<:Union{Number,Complex{<:Number}},
    T_f2<:Union{Number,Complex{<:Number}}
}

Composition of a combined symmetry group with a symmetry group acting on the same DoF-object. The resulting symmetry group combines the cycles, check functions, apply functions, and factors of the input symmetry groups.

Arguments

csg::SymBasis.SymGroups.SymGroup{B,T_s,T,Ti,T_f1}: The symmetry group.

sg::SymBasis.SymGroups.CombSymGroup{B,T_s,T,Ti,T_f2}: The combined symmetry group.

Returns

SymBasis.SymGroups.CombSymGroup{B,T_s,T,Ti,typeof(*(one(T_f1), one(T_f2)))}: The combined symmetry group.

Base.:∘ — Method

∘(
    sg1::SymGroup{B,T_s,T,Ti,T_f1},
    sg2::SymGroup{B,T_s,T,Ti,T_f2}
) where {
    B,
    T_s,
    T<:Integer,
    Ti<:Integer,
    T_f1<:Union{Number,Complex{<:Number}},
    T_f2<:Union{Number,Complex{<:Number}}
}

Composition of two symmetry groups acting on the same DoF-object. The resulting symmetry group combines the cycles, check functions, apply functions, and factors of the input symmetry groups.

Arguments

sg1::SymBasis.SymGroups.SymGroup{B,T_s,T,Ti,T_f1}: The first symmetry group.

sg2::SymBasis.SymGroups.SymGroup{B,T_s,T,Ti,T_f2}: The second symmetry group .

Returns

SymBasis.SymGroups.CombSymGroup{B,T_s,T,Ti,typeof(*(one(T_f1), one(T_f2)))}: The combined symmetry group.

SymBasis.SymGroups._check_Nₛ — Method

_check_Nₛ(
    state::SymBasis.DigitBase.BaseInt{T,Ti,B},
    p::NamedTuple{names,NT}
) where {T<:Integer,Ti<:Integer,B,names,NT<:Tuple{Vararg{Integer}}}

Internal function to check if the given state has the specified digit counts as defined in the named tuple p.

Arguments

state::SymBasis.DigitBase.BaseInt{T,Ti,B}: The state to be checked.

p::NamedTuple{names,NT}: A named tuple containing the digit counts.

Returns

Bool: true if the state has the specified digit counts, false otherwise.

SymBasis.SymGroups._check_Nₛ — Method

_check_Nₛ(
    state::SymBasis.DigitBase.BaseInt{T,Ti,2},
    p::NamedTuple{N0::TN, N1::TN, N::TN}
) where {T<:Integer,Ti<:Integer,TN<:Integer}

Internal function to check if the given binary state has the specified counts of 0s and 1s as defined in the named tuple p.

Arguments

state::SymBasis.DigitBase.BaseInt{T,Ti,2}: The binary state to be checked.

p::NamedTuple{N0::TN, N1::TN, N::TN}: A named tuple containing the counts of 0s and 1s.

Returns

Bool: true if the state has the specified counts of 0s and 1s, false otherwise.

SymBasis.SymGroups._make_hashset — Method

_make_hashset(csg::SymBasis.SymGroups.CombSymGroup)

Create a SymBasis.Miscs.SmallHashSet instance suitable for storing symmetry sector parameters based on the provided combined symmetry group csg.

Arguments

csg::SymBasis.SymGroups.CombSymGroup: The combined symmetry group for which to create the hash set.

SymBasis.SymGroups._make_hashset — Method

_make_hashset(sg::SymBasis.SymGroups.SymGroup)

Create a SymBasis.Miscs.SmallHashSet instance suitable for storing symmetry sector parameters based on the provided symmetry group sg.

Arguments

sg::SymBasis.SymGroups.SymGroup: The symmetry group for which to create the hash set.

Returns

SymBasis.Miscs.SmallHashSet: A new SymBasis.Miscs.SmallHashSet instance.

SymBasis.SymGroups.apply_Nₛ — Method

apply_Nₛ(
    p::NamedTuple{names,NT},
    state::SymBasis.DigitBase.BaseInt{T,Ti,B}
) where {names,NT<:Tuple{Vararg{Integer}},T<:Integer,Ti<:Integer,B}

Apply the symmetry operation defined by the digit counts in p to the given state. Since this is a symmetry where the state remains unchanged, the function simply returns the input state.

Arguments

p::NamedTuple{names,NT}: A named tuple containing the digit counts.

state::SymBasis.DigitBase.BaseInt{T,Ti,B}: The state to which the symmetry operation will be applied.

Returns

SymBasis.DigitBase.BaseInt{T,Ti,B}: The unchanged state.

SymBasis.SymGroups.apply_perm — Method

apply_perm(
    p::@NamedTuple{perm::Tperm},
    state::SymBasis.DigitBase.BaseInt{T,Ti,B}
) where {T<:Integer,Ti<:Integer,B,Tperm<:AbstractVector{<:Ti}}

Apply the permutation p.perm to the given state.

Arguments

p::@NamedTuple{perm::Tperm}: A named tuple containing the permutation.

state::SymBasis.DigitBase.BaseInt{T,Ti,B}: The state to which the permutation will be applied.

Returns

SymBasis.DigitBase.BaseInt{T,Ti,B}: The state after applying the permutation.

SymBasis.SymGroups.apply_perm — Method

apply_perm(
    p::@NamedTuple{perm::Tperm},
    state::SymBasis.DigitBase.BaseInt{T,Ti,2}
) where {
    T<:Integer,
    Ti<:Integer,
    Tperm<:BitPermutations.BitPermutation{T,<:BitPermutations.PermutationBackend{T}}
}

Apply the bit permutation p.perm to the given binary state in a more efficient way.

Arguments

p::@NamedTuple{perm::Tperm}: A named tuple containing the bit permutation.

state::SymBasis.DigitBase.BaseInt{T,Ti,2}: The binary state to which the bit permutation will be applied.

Returns

SymBasis.DigitBase.BaseInt{T,Ti,2}: The binary state after applying the bit permutation.

SymBasis.SymGroups.check_Nₛ — Method

check_Nₛ(
    p::NamedTuple{names,NT},
    state::SymBasis.DigitBase.BaseInt{T,Ti,B},
    prev_bool::Bool,
) where {names,NT<:Tuple{Vararg{Integer}},T<:Integer,Ti<:Integer,B}

Check if the given state has the specified digit counts as defined in the named tuple p. Since this is a symmetry check, the result is combined with prev_bool.

Arguments

p::NamedTuple{names,NT}: A named tuple containing the digit counts.

state::SymBasis.DigitBase.BaseInt{T,Ti,B}: The state to be checked.

prev_bool::Bool: The previous boolean value to be combined with the check result.

Returns

Bool: The combined result of the previous boolean and the digit count check.

SymBasis.SymGroups.check_perm — Method

check_perm(
    p::@NamedTuple{perm::Tperm},
    state::SymBasis.DigitBase.BaseInt{T,Ti,B},
    prev_bool::Bool
) where {
    T<:Integer,
    Ti<:Integer,
    B,
    Tperm<:Union{
        AbstractVector{<:Ti},
        BitPermutations.BitPermutation{T,<:BitPermutations.PermutationBackend{T}}
    }
}

Check if the given state is invariant under the permutation p.perm. Since permutations are symmetries, this function always returns prev_bool unchanged.

Arguments

p::@NamedTuple{perm::Tperm}: A named tuple containing the permutation.

state::SymBasis.DigitBase.BaseInt{T,Ti,B}: The state to be checked.

prev_bool::Bool: The previous boolean value to be combined with the check result.

Returns

Bool: The combined result of the previous boolean and the check (always prev_bool).

SymBasis.SymGroups.sym — Method

sym(s::Symbol, args...; kwargs...)

Create a symmetry group based on the specified symbol s and additional arguments.

Arguments

s::Symbol: The symbol representing the type of symmetry group to create.

args...: Additional positional arguments required for the specific symmetry group.

Keyword Arguments

kwargs...: Additional keyword arguments required for the specific symmetry group.

Returns

SymBasis.SymGroups.SymGroup: The created symmetry group.

SymBasis.SymGroups.sym — Method

sym(
    ::Val{:Translational},
    dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti},
    k::T_k,
    perm::AbstractVector{Ti}
) where {B,T_s,T<:Integer,Ti<:Integer,T_k<:Integer}

Create a translational symmetry group for the given DoF-object dofo, momentum k, and permutation perm.

Arguments

dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti}: The DoF-object.

k::T_k: The momentum number.

perm::AbstractVector{Ti}: The permutation vector defining the translation.

Returns

SymBasis.SymGroups.SymGroup: The translational symmetry group.

SymBasis.SymGroups.sym — Method

sym(
    ::Val{:SpatialReflection},
    dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti},
    p::T_p,
    perm::AbstractVector{Ti}
) where {B,T_s,T<:Integer,Ti<:Integer,T_p<:Integer}

Create a spatial reflection symmetry group for the given DoF-object dofo, parity p, and permutation perm.

Arguments

dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti}: The DoF-object.

p::T_p: The parity number (either -1 or 1).

perm::AbstractVector{Ti}: The permutation vector defining the spatial reflection.

Returns

SymBasis.SymGroups.SymGroup: The spatial reflection symmetry group.

SymBasis.SymGroups.sym — Method

sym(
    ::Val{:TotalMagnetization},
    dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti},
    mag::T_s,
    N::Integer
) where {B,T_s<:Rational,T<:Integer,Ti<:Integer}

Create a total magnetization symmetry group for the given spin object dofo, target magnetization mag, and number of sites N.

Arguments

dofo::SymBasis.DoFObjects.DoFObject{B,T_s,T,Ti}: The DoF-object representing the quantum mechanical spin.

mag::T_s: The target total magnetization.

N::Integer: The total number of DoF-objects in the system.

Returns

SymBasis.SymGroups.SymGroup: The total magnetization symmetry group.